requirements.txt
google-cloud-firestore==2.11.1
apache-airflow-providers-http==4.4.2
apache-airflow-providers-google==10.6.0